refactor: progress·instruction 도메인 내부 정리 및 패키지 구조 적용
[progress]
- util/util.py 신설: CONCEPT_KEY_BY_ANSWER, resolve_concept_key, infer_lesson_id 등
  service.py에 묻혀있던 순수 함수와 상수를 분리
- LearningClassroomRepository Protocol을 service.py → repository/repository.py로 이동
- router.py의 _calculate_progress_rate, _build_badges를
  service.py 메서드로 이동 (라우터는 HTTP 변환만 담당)

[instruction]
- validation/validation.py 신설: AI 문제 검증 순수 함수 5개 분리
- generation.py를 service/ 패키지 안으로 이동
- service.py의 StageProblemLookup 중복 정의 제거 (repository.py 단일화)

[공통]
- progress, instruction 모두 controller/dependency/repository/schema/service 패키지 구조로 전환
