+++ app/domains/classroom/controller/teacher_class_router.py
@@ -14,8 +14,8 @@ from app.domains.auth.dependencies import get_current_teacher
 from app.domains.auth.models import User
 from app.domains.classroom.dependency.dependencies import get_classroom_service
 from app.domains.classroom.service.service import ClassroomService
-from app.domains.progress.dependencies import get_learning_record_service
-from app.domains.progress.service import LearningRecordService
+from app.domains.progress.dependency.dependencies import get_learning_record_service
+from app.domains.progress.service.service import LearningRecordService
 from app.domains.classroom.schema.schemas import (
     AddStudentRequest,
     CreateClassRequest,
