+++ app/domains/content/stage1/router.py
@@ -9,8 +9,8 @@ from app.domains.content.stage1.schemas import (
     Stage1SubmitResponse,
 )
 from app.domains.content.stage1.service import stage1_pair_response
-from app.domains.progress.dependencies import get_learning_record_service
-from app.domains.progress.service import LearningRecordService
+from app.domains.progress.dependency.dependencies import get_learning_record_service
+from app.domains.progress.service.service import LearningRecordService
 from app.infrastructure.db.mongo.mongo_client import get_mongo_client
 
 router = APIRouter(prefix="/student/learning", tags=["student-learning"])
