+++ app/domains/instruction/service/service.py
@@ -10,8 +10,12 @@ from app.domains.instruction.models import (
     GeneratedProblem,
     TeacherAssignment,
 )
-from app.domains.progress.service import CONCEPT_KEY_BY_ANSWER
-from app.domains.instruction.repository import TeacherAssignmentRepository
+from app.domains.instruction.repository.repository import StageProblemLookup, TeacherAssignmentRepository
+from app.domains.instruction.validation.validation import (
+    allowed_concept_keys,
+    normalize_sentence,
+    validate_generated_problem,
+)
 
 
 class InstructionError(ValueError):
@@ -44,11 +48,6 @@ class ProblemGenerator(Protocol):
         lesson_id: str,
         difficulty: str | None = None,
     ) -> list[GeneratedProblem]:
-        ...
-
-
-class StageProblemLookup(Protocol):
-    def find_stage3_full_sentences(self, lesson_id: str) -> set[str]:
         ...
 
 
@@ -129,7 +128,7 @@ class InstructionService:
     ) -> tuple[TeacherAssignment, list[ProblemValidationResult]]:
         if stage != 3:
             raise InstructionError("현재 AI 문제 생성은 Stage 3만 지원합니다.")
-        if concept_key not in _allowed_concept_keys():
+        if concept_key not in allowed_concept_keys():
             raise InstructionError("지원하지 않는 concept_key입니다.")
 
         self._validate_target_access(
@@ -405,7 +404,7 @@ class InstructionService:
         seen_full_sentences: set[str] = set()
         results = []
         for problem in problems:
-            reasons = _validate_generated_problem(
+            reasons = validate_generated_problem(
                 problem=problem,
                 concept_key=concept_key,
                 existing_full_sentences=existing_full_sentences,
@@ -413,7 +412,7 @@ class InstructionService:
             )
             problem.validation_status = "invalid" if reasons else "valid"
             if not reasons:
-                seen_full_sentences.add(_normalize_sentence(problem.full_sentence))
+                seen_full_sentences.add(normalize_sentence(problem.full_sentence))
             results.append(
                 ProblemValidationResult(
                     problem=problem,
@@ -422,73 +421,3 @@ class InstructionService:
                 )
             )
         return results
-
-
-def _validate_generated_problem(
-    problem: GeneratedProblem,
-    concept_key: str,
-    existing_full_sentences: set[str],
-    seen_full_sentences: set[str],
-) -> list[str]:
-    reasons = []
-    if not problem.sentence_part1.strip() and not problem.sentence_part2.strip():
-        reasons.append("빈칸 앞/뒤 문장 중 하나는 필요합니다.")
-    if not problem.correct_answer.strip():
-        reasons.append("정답이 비어 있습니다.")
-    if not problem.full_sentence.strip():
-        reasons.append("완성 문장이 비어 있습니다.")
-    if len(problem.explanation.strip()) > 120:
-        reasons.append("해설은 120자 이하여야 합니다.")
-
-    allowed_answers = _answers_for_concept(concept_key)
-    if problem.correct_answer.strip() not in allowed_answers:
-        reasons.append("정답이 concept_key의 허용 답안 목록에 없습니다.")
-
-    expected_full_sentence = _compose_full_sentence(
-        problem.sentence_part1,
-        problem.correct_answer,
-        problem.sentence_part2,
-    )
-    if _normalize_sentence(problem.full_sentence) != _normalize_sentence(expected_full_sentence):
-        reasons.append("완성 문장이 빈칸 앞/정답/뒤 문장 조합과 일치하지 않습니다.")
-
-    normalized_full_sentence = _normalize_sentence(problem.full_sentence)
-    normalized_existing = {_normalize_sentence(sentence) for sentence in existing_full_sentences}
-    if normalized_full_sentence in normalized_existing:
-        reasons.append("기존 기본 문제와 중복됩니다.")
-    if normalized_full_sentence in seen_full_sentences:
-        reasons.append("같은 생성 요청 안에서 중복된 문제입니다.")
-    return reasons
-
-
-def _allowed_concept_keys() -> set[str]:
-    return set(CONCEPT_KEY_BY_ANSWER.values())
-
-
-def _answers_for_concept(concept_key: str) -> set[str]:
-    return {
-        answer
-        for answer, mapped_concept_key in CONCEPT_KEY_BY_ANSWER.items()
-        if mapped_concept_key == concept_key
-    }
-
-
-def _compose_full_sentence(part1: str, answer: str, part2: str) -> str:
-    left = part1.strip()
-    middle = answer.strip()
-    right = part2.strip()
-
-    if left:
-        sentence = f"{left} {middle}"
-    else:
-        sentence = middle
-
-    if not right:
-        return sentence
-    if right[0] in ".,?!)]}":
-        return f"{sentence}{right}"
-    return f"{sentence} {right}"
-
-
-def _normalize_sentence(sentence: str) -> str:
-    return " ".join(sentence.strip().split())
