+++ app/domains/progress/service/service.py
@@ -1,6 +1,6 @@
 from collections import defaultdict
-from datetime import datetime, timezone
-from typing import Any, Dict, Optional, Protocol
+from datetime import datetime
+from typing import Dict, Optional
 
 from app.common.security import utc_now
 from app.domains.progress.models import (
@@ -8,52 +8,14 @@ from app.domains.progress.models import (
     StudentWeaknessProfile,
     WeakConcept,
 )
-from app.domains.progress.repository import LearningRecordRepository
-
-
-CONCEPT_KEY_BY_ANSWER = {
-    # Stage 2/3 활용형
-    "가르쳐": "가르치다/가르키다",
-    "가르켰다": "가르치다/가르키다",
-    "맞췄다": "맞추다/맞히다",
-    "맞혀": "맞추다/맞히다",
-    "맞혔다": "맞추다/맞히다",
-    "잃어버렸다": "잊다/잃다",
-    "잊었다": "잊다/잃다",
-    "메고": "메다/매다",
-    "매고": "메다/매다",
-    "매었다": "메다/매다",
-    "바란다": "바라다/바래다",
-    "바랬다": "바라다/바래다",
-    "부쳤다": "부치다/붙이다",
-    "붙였다": "부치다/붙이다",
-    "부치신다": "부치다/붙이다",
-    "돼": "되/돼",
-    "되고": "되/돼",
-    "돼서": "되/돼",
-    "안": "안/않다",
-    "않다": "안/않다",
-    "않고": "안/않다",
-    "반드시": "반드시/반듯이",
-    "반듯이": "반드시/반듯이",
-    "이따가": "이따가/있다가",
-    "있다가": "이따가/있다가",
-    # Stage 1 기본형 (카드 쌍 word1/word2)
-    "가르치다": "가르치다/가르키다",
-    "가르키다": "가르치다/가르키다",
-    "맞추다": "맞추다/맞히다",
-    "맞히다": "맞추다/맞히다",
-    "잊다": "잊다/잃다",
-    "잃다": "잊다/잃다",
-    "메다": "메다/매다",
-    "매다": "메다/매다",
-    "바라다": "바라다/바래다",
-    "바래다": "바라다/바래다",
-    "부치다": "부치다/붙이다",
-    "붙이다": "부치다/붙이다",
-    "되다": "되/돼",
-    "돼다": "되/돼",
-}
+from app.domains.progress.repository.repository import LearningClassroomRepository, LearningRecordRepository
+from app.domains.progress.util.util import (
+    _calculate_priority,
+    _record_created_at,
+    build_problem_key,
+    infer_lesson_id,
+    resolve_concept_key,
+)
 
 
 class LearningRecordService:
@@ -126,11 +88,9 @@ class LearningRecordService:
         for record in records:
             if record.get("is_correct"):
                 continue
-
             concept_key = record.get("concept_key")
             if not concept_key:
                 continue
-
             wrong_count_by_concept[concept_key] += 1
             created_at = record.get("created_at")
             if isinstance(created_at, datetime):
@@ -192,6 +152,32 @@ class LearningRecordService:
             "completed_question_count": len(completed_question_ids),
         }
 
+    def calculate_progress_rate(self, user_id: str, units_with_lessons: list) -> int:
+        """Stage 2를 통과한 차시 수 / 전체 차시 수로 진행도를 계산한다."""
+        try:
+            total_lessons = sum(len(lessons) for _, lessons in units_with_lessons)
+            if total_lessons == 0:
+                return 0
+            records = self.get_records(user_id)
+            completed_lesson_ids = {
+                record.lesson_id
+                for record in records
+                if record.is_correct and record.lesson_id and record.stage == 2
+            }
+            return round(len(completed_lesson_ids) / total_lessons * 100)
+        except Exception:
+            return 0
+
+    def build_progress_badges(self, metrics: dict[str, int], progress_rate: int) -> list[str]:
+        badges = []
+        if metrics["total_solved_count"] > 0:
+            badges.append("첫 학습 시작")
+        if metrics["streak_correct_count"] >= 3:
+            badges.append("연속 정답")
+        if progress_rate >= 100:
+            badges.append("전체 학습 완료")
+        return badges
+
     def record_stage1_card_check(
         self,
         pair_id: str,
@@ -199,10 +185,6 @@ class LearningRecordService:
         chosen_word: str,
         user_id: Optional[str] = None,
     ) -> tuple[bool, str]:
-        """1단계 카드 확인 결과를 평가하고, 로그인 사용자면 기록을 저장한다.
-
-        Returns: (is_correct, concept_key)
-        """
         is_correct = chosen_word.strip() == correct_word.strip()
         concept_key = resolve_concept_key(correct_word, chosen_word)
         if user_id:
@@ -226,10 +208,6 @@ class LearningRecordService:
         user_answer: str,
         user_id: Optional[str] = None,
     ) -> tuple[bool, str]:
-        """2단계 문제 답안을 평가하고, 로그인 사용자면 기록을 저장한다.
-
-        Returns: (is_correct, concept_key)
-        """
         is_correct = user_answer.strip() == correct_answer.strip()
         concept_key = resolve_concept_key(correct_answer, user_answer)
         if user_id:
@@ -251,86 +229,14 @@ class LearningRecordService:
             return 1
         records = self.repository.find_records_by_user(user_id)
         previous_attempts = sum(
-            1
-            for record in records
-            if record.get("problem_key") == problem_key
+            1 for record in records if record.get("problem_key") == problem_key
         )
         return previous_attempts + 1
 
     def _resolve_class_id(self, user_id: str) -> str | None:
         if self.classroom_repository is None:
             return None
-
         classrooms = self.classroom_repository.find_classes_by_student(user_id)
         if not classrooms:
             return None
         return classrooms[0].get("class_id")
-
-
-class LearningClassroomRepository(Protocol):
-    def find_classes_by_student(self, student_id: str) -> list[dict[str, Any]]:
-        ...
-
-
-def resolve_concept_key(correct_answer: str, user_answer: str | None = None) -> str:
-    correct = correct_answer.strip()
-    submitted = (user_answer or "").strip()
-    return (
-        CONCEPT_KEY_BY_ANSWER.get(correct)
-        or CONCEPT_KEY_BY_ANSWER.get(submitted)
-        or correct
-    )
-
-
-def build_problem_key(
-    stage: int,
-    lesson_id: str | None,
-    problem_id: str | int | None,
-) -> str | None:
-    if lesson_id is None or problem_id is None:
-        return None
-    return f"stage{stage}:{lesson_id}:{problem_id}"
-
-
-def infer_lesson_id(stage: int, problem_id: str | int | None) -> str | None:
-    if problem_id is None:
-        return None
-
-    if stage == 1:
-        pair_no = _extract_last_int(problem_id)
-        if pair_no is None:
-            return None
-        return f"lesson_{((pair_no - 1) // 2) + 1}"
-
-    numeric_problem_id = _extract_last_int(problem_id)
-    if numeric_problem_id is None:
-        return None
-
-    if stage == 2:
-        return f"lesson_{((numeric_problem_id - 1) // 4) + 1}"
-    if stage == 3:
-        return f"lesson_{((numeric_problem_id - 1) // 5) + 1}"
-    return None
-
-
-def _extract_last_int(value: str | int) -> int | None:
-    if isinstance(value, int):
-        return value
-    parts = str(value).split("_")
-    for part in reversed(parts):
-        if part.isdigit():
-            return int(part)
-    return None
-
-
-def _calculate_priority(wrong_count: int) -> float:
-    return min(1.0, round(0.5 + wrong_count * 0.15, 2))
-
-
-def _record_created_at(record: Dict[str, Any]) -> datetime:
-    created_at = record.get("created_at")
-    if not isinstance(created_at, datetime):
-        return datetime.min.replace(tzinfo=timezone.utc)
-    if created_at.tzinfo is None:
-        return created_at.replace(tzinfo=timezone.utc)
-    return created_at.astimezone(timezone.utc)
