+++ tests/learning/test_record_service.py
@@ -1,11 +1,7 @@
 from datetime import datetime, timedelta, timezone
 
-from app.domains.progress.service import (
-    LearningRecordService,
-    build_problem_key,
-    infer_lesson_id,
-    resolve_concept_key,
-)
+from app.domains.progress.service.service import LearningRecordService
+from app.domains.progress.util.util import build_problem_key, infer_lesson_id, resolve_concept_key
 
 
 class FakeLearningRecordRepository:
