+++ tests/structure/test_domain_structure.py
@@ -16,7 +16,7 @@ def test_domain_packages_are_importable():
         "app.domains.auth.service",
         "app.domains.classroom.controller.teacher_class_router",
         "app.domains.classroom.controller.student_class_router",
-        "app.domains.progress.teacher_router",
+        "app.domains.progress.controller.teacher_router",
         "app.domains.classroom.dependency.dependencies",
         "app.domains.classroom.models",
         "app.domains.classroom.repository.repository",
@@ -24,13 +24,14 @@ def test_domain_packages_are_importable():
         "app.domains.classroom.schema.schemas",
         "app.domains.agent.schemas",
         "app.domains.agent.service",
-        "app.domains.instruction.student_router",
-        "app.domains.instruction.teacher_router",
-        "app.domains.instruction.generation",
+        "app.domains.instruction.controller.student_router",
+        "app.domains.instruction.controller.teacher_router",
+        "app.domains.instruction.service.generation",
+        "app.domains.instruction.validation.validation",
         "app.domains.instruction.models",
-        "app.domains.instruction.repository",
-        "app.domains.instruction.schemas",
-        "app.domains.instruction.service",
+        "app.domains.instruction.repository.repository",
+        "app.domains.instruction.schema.schemas",
+        "app.domains.instruction.service.service",
         "app.domains.content.controller.router",
         "app.domains.content.stage1.router",
         "app.domains.content.stage1.schemas",
@@ -38,7 +39,7 @@ def test_domain_packages_are_importable():
         "app.domains.content.stage2.router",
         "app.domains.content.stage2.schemas",
         "app.domains.content.stage2.service",
-        "app.domains.progress.router",
+        "app.domains.progress.controller.router",
         "app.domains.content.stage3.router",
         "app.domains.content",
         "app.domains.content.models",
@@ -46,9 +47,9 @@ def test_domain_packages_are_importable():
         "app.domains.content.schema.schemas",
         "app.domains.content.service.service",
         "app.domains.progress.models",
-        "app.domains.progress.repository",
-        "app.domains.progress.schemas",
-        "app.domains.progress.service",
+        "app.domains.progress.repository.repository",
+        "app.domains.progress.schema.schemas",
+        "app.domains.progress.service.service",
         "app.domains.content.stage3.schemas",
         "app.domains.content.stage3.service",
         "app.domains.agent",
